Move expectBars to its own file
so that I can reuse it in other test files later
